requirements.txt
numpy==2.0.0
Pillow==10.4.0
pygame==2.5.2
pygame==2.6.0
pynput==1.7.7
PySide6==6.7.2
PySide6==6.7.2
PySide6_Addons==6.7.1
PySide6_Addons==6.7.2
PySide6_Essentials==6.7.1
PySide6_Essentials==6.7.2
pywin32==306
